requirements.txt
anthropic>=0.18
pillow>=10.0
pyaudioop